backend/requirements.txt
# Constitutional AI - Backend Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & RAG
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
openai==1.10.0
tiktoken==0.5.2

# Vector Database
pinecone-client==5.0.1
# Alternative: weaviate-client==3.26.0

# Embeddings
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# Search
elasticsearch==8.11.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
